fix(asset-inventory): refactor lsb
