+++ apps/web/src/services/asset-inventory/AssetInventoryContainer.vue
@@ -4,20 +4,15 @@ import {
 } from 'vue';
 import { useRoute } from 'vue-router/composables';
 
-import { clone } from 'lodash';
-
 import { LocalStorageAccessor } from '@cloudforet/core-lib/local-storage-accessor';
 
 import { store } from '@/store';
-
-import { MENU_ID } from '@/lib/menu/config';
 
 import CenteredPageLayout from '@/common/modules/page-layouts/CenteredPageLayout.vue';
 import GeneralPageLayout from '@/common/modules/page-layouts/GeneralPageLayout.vue';
 import VerticalPageLayout from '@/common/modules/page-layouts/VerticalPageLayout.vue';
 
-import CloudServiceLSB from '@/services/asset-inventory/components/CloudServiceLSB.vue';
-import SecurityLSB from '@/services/asset-inventory/components/SecurityLSB.vue';
+import AssetInventoryLSB from '@/services/asset-inventory/AssetInventoryLSB.vue';
 import { useAssetInventorySettingsStore } from '@/services/asset-inventory/stores/asset-inventory-settings-store';
 import { useCloudServiceDetailPageStore } from '@/services/asset-inventory/stores/cloud-service-detail-page-store';
 import { useCloudServicePageStore } from '@/services/asset-inventory/stores/cloud-service-page-store';
@@ -33,11 +28,6 @@ const storeState = reactive({
 });
 const state = reactive({
     lsbVisible: computed<boolean>(() => route.meta?.lsbVisible),
-    menuId: computed<string>(() => {
-        const reversedMatched = clone(route.matched).reverse();
-        const closestRoute = reversedMatched.find((d) => d.meta?.menuId !== undefined);
-        return closestRoute?.meta?.menuId;
-    }),
 });
 assetInventorySettings.initState();
 assetInventorySettings.$onAction((action) => {
@@ -62,8 +52,7 @@ onUnmounted(() => {
     <fragment>
         <vertical-page-layout v-if="state.lsbVisible">
             <template #sidebar>
-                <cloud-service-l-s-b v-if="state.menuId === MENU_ID.CLOUD_SERVICE" />
-                <security-l-s-b v-if="state.menuId === MENU_ID.SECURITY" />
+                <asset-inventory-l-s-b />
             </template>
             <template #default>
                 <router-view />
